+++ integrations/pgvector/src/haystack_integrations/document_stores/pgvector/document_store.py
@@ -4,7 +4,7 @@
 from typing import Any, Dict, List, Literal, Optional
 
 from haystack import default_from_dict, default_to_dict, logging
-from haystack.dataclasses.document import ByteStream, Document
+from haystack.dataclasses.document import Document
 from haystack.document_stores.errors import DocumentStoreError, DuplicateDocumentError
 from haystack.document_stores.types import DuplicatePolicy
 from haystack.utils.auth import Secret, deserialize_secrets_inplace
@@ -15,11 +15,11 @@ from psycopg.cursor import Cursor
 from psycopg.rows import dict_row
 from psycopg.sql import SQL, Identifier
 from psycopg.sql import Literal as SQLLiteral
-from psycopg.types.json import Jsonb
 
 from pgvector.psycopg import register_vector, register_vector_async
 
-from .filters import _convert_filters_to_where_clause_and_params
+from .converters import _from_haystack_to_pg_documents, _from_pg_to_haystack_documents
+from .filters import _convert_filters_to_where_clause_and_params, _validate_filters
 
 logger = logging.getLogger(__name__)
 
@@ -166,123 +166,6 @@ class PgvectorDocumentStore:
         self._async_dict_cursor = None
         self._table_initialized = False
 
-    def _ensure_valid_connection(self):
-        """
-        Internal method to ensure that the connection to the PostgreSQL database exists and is valid.
-        If not, it will be created and the cursors will be initialized.
-        """
-        if self._connection and self._cursor and self._dict_cursor and self._connection_is_valid(self._connection):
-            return
-
-        # close the connection if it already exists
-        if self._connection:
-            try:
-                self._connection.close()
-            except Error as e:
-                logger.debug("Failed to close connection: {e}", e=str(e))
-
-        conn_str = self.connection_string.resolve_value() or ""
-        connection = Connection.connect(conn_str)
-        connection.autocommit = True
-        if self.create_extension:
-            connection.execute("CREATE EXTENSION IF NOT EXISTS vector")
-        register_vector(connection)  # Note: this must be called before creating the cursors.
-
-        self._connection = connection
-        self._cursor = self._connection.cursor()
-        self._dict_cursor = self._connection.cursor(row_factory=dict_row)
-
-        if not self._table_initialized:
-            self._initialize_table()
-
-    async def _ensure_valid_connection_async(self):
-        """
-        Internal method to ensure that the async connection to the PostgreSQL database exists and is valid.
-        If not, it will be created and the cursors will be initialized.
-        """
-
-        if (
-            self._async_connection
-            and self._async_cursor
-            and self._async_dict_cursor
-            and await self._connection_is_valid_async(self._async_connection)
-        ):
-            return
-
-        # close the connection if it already exists
-        if self._async_connection:
-            await self._async_connection.close()
-
-        conn_str = self.connection_string.resolve_value() or ""
-        async_connection = await AsyncConnection.connect(conn_str)
-        await async_connection.set_autocommit(True)
-        if self.create_extension:
-            await async_connection.execute("CREATE EXTENSION IF NOT EXISTS vector")
-        await register_vector_async(async_connection)  # Note: this must be called before creating the cursors.
-
-        self._async_connection = async_connection
-        self._async_cursor = self._async_connection.cursor()
-        self._async_dict_cursor = self._async_connection.cursor(row_factory=dict_row)
-
-        if not self._table_initialized:
-            await self._initialize_table_async()
-
-    def _initialize_table(self):
-        """
-        Internal method to initialize the table.
-        """
-        if self.recreate_table:
-            self.delete_table()
-
-        self._create_table_if_not_exists()
-        self._create_keyword_index_if_not_exists()
-
-        if self.search_strategy == "hnsw":
-            self._handle_hnsw()
-
-        self._table_initialized = True
-
-    async def _initialize_table_async(self):
-        """
-        Internal async method to initialize the table.
-        """
-        if self.recreate_table:
-            await self.delete_table_async()
-
-        await self._create_table_if_not_exists_async()
-        await self._create_keyword_index_if_not_exists_async()
-
-        if self.search_strategy == "hnsw":
-            await self._handle_hnsw_async()
-
-        self._table_initialized = True
-
-    @staticmethod
-    def _connection_is_valid(connection):
-        """
-        Internal method to check if the connection is still valid.
-        """
-
-        # implementation inspired to psycopg pool
-        # https://github.com/psycopg/psycopg/blob/d38cf7798b0c602ff43dac9f20bbab96237a9c38/psycopg_pool/psycopg_pool/pool.py#L528
-
-        try:
-            connection.execute("")
-        except Error:
-            return False
-        return True
-
-    @staticmethod
-    async def _connection_is_valid_async(connection):
-        """
-        Internal method to check if the async connection is still valid.
-        """
-        try:
-            await connection.execute("")
-        except Error:
-            return False
-        return True
-
     def to_dict(self) -> Dict[str, Any]:
         """
         Serializes the component to a dictionary.
@@ -320,6 +203,32 @@ class PgvectorDocumentStore:
         """
         deserialize_secrets_inplace(data["init_parameters"], ["connection_string"])
         return default_from_dict(cls, data)
+
+    @staticmethod
+    def _connection_is_valid(connection):
+        """
+        Internal method to check if the connection is still valid.
+        """
+
+        # implementation inspired to psycopg pool
+        # https://github.com/psycopg/psycopg/blob/d38cf7798b0c602ff43dac9f20bbab96237a9c38/psycopg_pool/psycopg_pool/pool.py#L528
+
+        try:
+            connection.execute("")
+        except Error:
+            return False
+        return True
+
+    @staticmethod
+    async def _connection_is_valid_async(connection):
+        """
+        Internal method to check if the async connection is still valid.
+        """
+        try:
+            await connection.execute("")
+        except Error:
+            return False
+        return True
 
     def _execute_sql(
         self, sql_query: Query, params: Optional[tuple] = None, error_msg: str = "", cursor: Optional[Cursor] = None
@@ -379,49 +288,162 @@ class PgvectorDocumentStore:
 
         return result
 
-    def _create_table_if_not_exists(self):
+    def _ensure_db_setup(self):
         """
-        Creates the table to store Haystack documents if it doesn't exist yet.
+        Ensures that the connection to the PostgreSQL database exists and is valid.
+        If not, connection and cursors are created.
+        If the table is not initialized, it will be set up.
         """
+        if self._connection and self._cursor and self._dict_cursor and self._connection_is_valid(self._connection):
+            return
+
+        # close the connection if it already exists
+        if self._connection:
+            try:
+                self._connection.close()
+            except Error as e:
+                logger.debug("Failed to close connection: {e}", e=str(e))
+
+        conn_str = self.connection_string.resolve_value() or ""
+        connection = Connection.connect(conn_str)
+        connection.autocommit = True
+        if self.create_extension:
+            connection.execute("CREATE EXTENSION IF NOT EXISTS vector")
+        register_vector(connection)  # Note: this must be called before creating the cursors.
+
+        self._connection = connection
+        self._cursor = self._connection.cursor()
+        self._dict_cursor = self._connection.cursor(row_factory=dict_row)
+
+        if not self._table_initialized:
+            self._initialize_table()
+
+    async def _ensure_db_setup_async(self):
+        """
+        Async internal method.
+        Ensures that the connection to the PostgreSQL database exists and is valid.
+        If not, connection and cursors are created.
+        If the table is not initialized, it will be set up.
+        """
+
+        if (
+            self._async_connection
+            and self._async_cursor
+            and self._async_dict_cursor
+            and await self._connection_is_valid_async(self._async_connection)
+        ):
+            return
+
+        # close the connection if it already exists
+        if self._async_connection:
+            await self._async_connection.close()
+
+        conn_str = self.connection_string.resolve_value() or ""
+        async_connection = await AsyncConnection.connect(conn_str)
+        await async_connection.set_autocommit(True)
+        if self.create_extension:
+            await async_connection.execute("CREATE EXTENSION IF NOT EXISTS vector")
+        await register_vector_async(async_connection)  # Note: this must be called before creating the cursors.
+
+        self._async_connection = async_connection
+        self._async_cursor = self._async_connection.cursor()
+        self._async_dict_cursor = self._async_connection.cursor(row_factory=dict_row)
+
+        if not self._table_initialized:
+            await self._initialize_table_async()
+
+    def _build_table_creation_queries(self):
+        """
+        Internal method to build the SQL queries for table creation.
+        """
+
+        sql_table_exists = SQL("SELECT 1 FROM pg_tables WHERE schemaname = %s AND tablename = %s")
+        sql_create_table = SQL(CREATE_TABLE_STATEMENT).format(
+            schema_name=Identifier(self.schema_name),
+            table_name=Identifier(self.table_name),
+            embedding_dimension=SQLLiteral(self.embedding_dimension),
+        )
+
+        sql_keyword_index_exists = SQL(
+            "SELECT 1 FROM pg_indexes WHERE schemaname = %s AND tablename = %s AND indexname = %s"
+        )
+        sql_create_keyword_index = SQL(
+            "CREATE INDEX {index_name} ON {schema_name}.{table_name} USING GIN (to_tsvector({language}, content))"
+        ).format(
+            schema_name=Identifier(self.schema_name),
+            index_name=Identifier(self.keyword_index_name),
+            table_name=Identifier(self.table_name),
+            language=SQLLiteral(self.language),
+        )
+
+        return sql_table_exists, sql_create_table, sql_keyword_index_exists, sql_create_keyword_index
+
+    def _initialize_table(self):
+        """
+        Internal method to initialize the table.
+        """
+        if self.recreate_table:
+            self.delete_table()
+
+        sql_table_exists, sql_create_table, sql_keyword_index_exists, sql_create_keyword_index = (
+            self._build_table_creation_queries()
+        )
 
         table_exists = bool(
             self._execute_sql(
-                "SELECT 1 FROM pg_tables WHERE schemaname = %s AND tablename = %s",
-                (self.schema_name, self.table_name),
-                "Could not check if table exists",
+                sql_table_exists, (self.schema_name, self.table_name), "Could not check if table exists"
             ).fetchone()
         )
-
         if not table_exists:
-            create_sql = SQL(CREATE_TABLE_STATEMENT).format(
-                schema_name=Identifier(self.schema_name),
-                table_name=Identifier(self.table_name),
-                embedding_dimension=SQLLiteral(self.embedding_dimension),
-            )
+            self._execute_sql(sql_create_table, error_msg="Could not create table")
 
-            self._execute_sql(create_sql, error_msg="Could not create table in PgvectorDocumentStore")
+        index_exists = bool(
+            self._execute_sql(
+                sql_keyword_index_exists,
+                (self.schema_name, self.table_name, self.keyword_index_name),
+                "Could not check if keyword index exists",
+            ).fetchone()
+        )
+        if not index_exists:
+            self._execute_sql(sql_create_keyword_index, error_msg="Could not create keyword index on table")
 
-    async def _create_table_if_not_exists_async(self):
+        if self.search_strategy == "hnsw":
+            self._handle_hnsw()
+
+        self._table_initialized = True
+
+    async def _initialize_table_async(self):
         """
-        Internal async method to create the table if it doesn't exist.
+        Internal async method to initialize the table.
         """
-        await self._execute_sql_async(
-            "SELECT 1 FROM pg_tables WHERE schemaname = %s AND tablename = %s",
-            (self.schema_name, self.table_name),
-            "Could not check if table exists",
-            self._async_cursor,
+        if self.recreate_table:
+            await self.delete_table_async()
+
+        sql_table_exists, sql_create_table, sql_keyword_index_exists, sql_create_keyword_index = (
+            self._build_table_creation_queries()
         )
 
+        await self._execute_sql_async(
+            sql_table_exists, (self.schema_name, self.table_name), "Could not check if table exists", self._async_cursor
+        )
         table_exists = bool(await self._async_cursor.fetchone())
-
         if not table_exists:
-            create_sql = SQL(CREATE_TABLE_STATEMENT).format(
-                schema_name=Identifier(self.schema_name),
-                table_name=Identifier(self.table_name),
-                embedding_dimension=SQLLiteral(self.embedding_dimension),
-            )
+            await self._execute_sql_async(sql_create_table, error_msg="Could not create table")
 
-            await self._execute_sql_async(create_sql, error_msg="Could not create table in PgvectorDocumentStore")
+        await self._execute_sql_async(
+            sql_keyword_index_exists,
+            (self.schema_name, self.table_name, self.keyword_index_name),
+            "Could not check if keyword index exists",
+            self._async_cursor,
+        )
+        index_exists = bool(await self._async_cursor.fetchone())
+        if not index_exists:
+            await self._execute_sql_async(sql_create_keyword_index, error_msg="Could not create keyword index on table")
+
+        if self.search_strategy == "hnsw":
+            await self._handle_hnsw_async()
+
+        self._table_initialized = True
 
     def delete_table(self):
         """
@@ -452,55 +474,6 @@ class PgvectorDocumentStore:
             delete_sql,
             error_msg=f"Could not delete table {self.schema_name}.{self.table_name} in PgvectorDocumentStore",
         )
-
-    def _create_keyword_index_if_not_exists(self):
-        """
-        Internal method to create the keyword index if not exists.
-        """
-        index_exists = bool(
-            self._execute_sql(
-                "SELECT 1 FROM pg_indexes WHERE schemaname = %s AND tablename = %s AND indexname = %s",
-                (self.schema_name, self.table_name, self.keyword_index_name),
-                "Could not check if keyword index exists",
-            ).fetchone()
-        )
-
-        if not index_exists:
-            sql_create_index = SQL(
-                "CREATE INDEX {index_name} ON {schema_name}.{table_name} USING GIN (to_tsvector({language}, content))"
-            ).format(
-                schema_name=Identifier(self.schema_name),
-                index_name=Identifier(self.keyword_index_name),
-                table_name=Identifier(self.table_name),
-                language=SQLLiteral(self.language),
-            )
-            self._execute_sql(sql_create_index, error_msg="Could not create keyword index on table")
-
-    async def _create_keyword_index_if_not_exists_async(self):
-        """
-        Internal async method to create the keyword index if not exists.
-        """
-
-        await self._execute_sql_async(
-            "SELECT 1 FROM pg_indexes WHERE schemaname = %s AND tablename = %s AND indexname = %s",
-            (self.schema_name, self.table_name, self.keyword_index_name),
-            "Could not check if keyword index exists",
-            self._async_cursor,
-        )
-
-        index_exists = bool(await self._async_cursor.fetchone())
-
-        if not index_exists:
-            sql_create_index = SQL(
-                "CREATE INDEX {index_name} ON {schema_name}.{table_name} USING GIN (to_tsvector({language}, content))"
-            ).format(
-                schema_name=Identifier(self.schema_name),
-                index_name=Identifier(self.keyword_index_name),
-                table_name=Identifier(self.table_name),
-                language=SQLLiteral(self.language),
-            )
-
-            await self._execute_sql_async(sql_create_index, error_msg="Could not create keyword index on table")
 
     def _build_hnsw_queries(self):
         """Common method to build all HNSW-related SQL queries"""
@@ -615,7 +588,7 @@ class PgvectorDocumentStore:
             schema_name=Identifier(self.schema_name), table_name=Identifier(self.table_name)
         )
 
-        self._ensure_valid_connection()
+        self._ensure_db_setup()
         count = self._execute_sql(sql_count, error_msg="Could not count documents in PgvectorDocumentStore").fetchone()[
             0
         ]
@@ -629,7 +602,7 @@ class PgvectorDocumentStore:
             schema_name=Identifier(self.schema_name), table_name=Identifier(self.table_name)
         )
 
-        await self._ensure_valid_connection_async()
+        await self._ensure_db_setup_async()
 
         await self._execute_sql_async(
             sql_count, error_msg="Could not count documents in PgvectorDocumentStore", cursor=self._async_cursor
@@ -649,13 +622,7 @@ class PgvectorDocumentStore:
         :raises TypeError: If `filters` is not a dictionary.
         :returns: A list of Documents that match the given filters.
         """
-        if filters:
-            if not isinstance(filters, dict):
-                msg = "Filters must be a dictionary"
-                raise TypeError(msg)
-            if "operator" not in filters and "conditions" not in filters:
-                msg = "Invalid filter syntax. See https://docs.haystack.deepset.ai/docs/metadata-filtering for details."
-                raise ValueError(msg)
+        _validate_filters(filters)
 
         sql_filter = SQL("SELECT * FROM {schema_name}.{table_name}").format(
             schema_name=Identifier(self.schema_name), table_name=Identifier(self.table_name)
@@ -666,7 +633,7 @@ class PgvectorDocumentStore:
             sql_where_clause, params = _convert_filters_to_where_clause_and_params(filters)
             sql_filter += sql_where_clause
 
-        self._ensure_valid_connection()
+        self._ensure_db_setup()
         result = self._execute_sql(
             sql_filter,
             params,
@@ -675,20 +642,14 @@ class PgvectorDocumentStore:
         )
 
         records = result.fetchall()
-        docs = self._from_pg_to_haystack_documents(records)
+        docs = _from_pg_to_haystack_documents(records)
         return docs
 
     async def filter_documents_async(self, filters: Optional[Dict[str, Any]] = None) -> List[Document]:
         """
         Returns the documents that match the filters provided.
         """
-        if filters:
-            if not isinstance(filters, dict):
-                msg = "Filters must be a dictionary"
-                raise TypeError(msg)
-            if "operator" not in filters and "conditions" not in filters:
-                msg = "Invalid filter syntax. See https://docs.haystack.deepset.ai/docs/metadata-filtering for details."
-                raise ValueError(msg)
+        _validate_filters(filters)
 
         sql_filter = SQL("SELECT * FROM {schema_name}.{table_name}").format(
             schema_name=Identifier(self.schema_name), table_name=Identifier(self.table_name)
@@ -699,7 +660,7 @@ class PgvectorDocumentStore:
             sql_where_clause, params = _convert_filters_to_where_clause_and_params(filters)
             sql_filter += sql_where_clause
 
-        await self._ensure_valid_connection_async()
+        await self._ensure_db_setup_async()
         result = await self._execute_sql_async(
             sql_filter,
             params,
@@ -708,8 +669,25 @@ class PgvectorDocumentStore:
         )
 
         records = await result.fetchall()
-        docs = self._from_pg_to_haystack_documents(records)
+        docs = _from_pg_to_haystack_documents(records)
         return docs
+
+    def _build_insert_statement(self, policy: DuplicatePolicy):
+        """
+        Builds the SQL insert statement to write documents.
+        """
+        sql_insert = SQL(INSERT_STATEMENT).format(
+            schema_name=Identifier(self.schema_name), table_name=Identifier(self.table_name)
+        )
+
+        if policy == DuplicatePolicy.OVERWRITE:
+            sql_insert += SQL(UPDATE_STATEMENT)
+        elif policy == DuplicatePolicy.SKIP:
+            sql_insert += SQL("ON CONFLICT DO NOTHING")
+
+        sql_insert += SQL(" RETURNING id")
+
+        return sql_insert
 
     def write_documents(self, documents: List[Document], policy: DuplicatePolicy = DuplicatePolicy.NONE) -> int:
         """
@@ -729,20 +707,11 @@ class PgvectorDocumentStore:
         if policy == DuplicatePolicy.NONE:
             policy = DuplicatePolicy.FAIL
 
-        db_documents = self._from_haystack_to_pg_documents(documents)
+        db_documents = _from_haystack_to_pg_documents(documents)
 
-        sql_insert = SQL(INSERT_STATEMENT).format(
-            schema_name=Identifier(self.schema_name), table_name=Identifier(self.table_name)
-        )
+        sql_insert = self._build_insert_statement(policy)
 
-        if policy == DuplicatePolicy.OVERWRITE:
-            sql_insert += SQL(UPDATE_STATEMENT)
-        elif policy == DuplicatePolicy.SKIP:
-            sql_insert += SQL("ON CONFLICT DO NOTHING")
-
-        sql_insert += SQL(" RETURNING id")
-
-        self._ensure_valid_connection()
+        self._ensure_db_setup()
         sql_query_str = sql_insert.as_string(self._cursor) if not isinstance(sql_insert, str) else sql_insert
         logger.debug("SQL query: {query}\nParameters: {parameters}", query=sql_query_str, parameters=db_documents)
 
@@ -784,20 +753,11 @@ class PgvectorDocumentStore:
         if policy == DuplicatePolicy.NONE:
             policy = DuplicatePolicy.FAIL
 
-        db_documents = self._from_haystack_to_pg_documents(documents)
+        db_documents = _from_haystack_to_pg_documents(documents)
 
-        sql_insert = SQL(INSERT_STATEMENT).format(
-            schema_name=Identifier(self.schema_name), table_name=Identifier(self.table_name)
-        )
+        sql_insert = self._build_insert_statement(policy)
 
-        if policy == DuplicatePolicy.OVERWRITE:
-            sql_insert += SQL(UPDATE_STATEMENT)
-        elif policy == DuplicatePolicy.SKIP:
-            sql_insert += SQL("ON CONFLICT DO NOTHING")
-
-        sql_insert += SQL(" RETURNING id")
-
-        await self._ensure_valid_connection_async()
+        await self._ensure_db_setup_async()
         sql_query_str = sql_insert.as_string(self._async_cursor) if not isinstance(sql_insert, str) else sql_insert
         logger.debug("SQL query: {query}\nParameters: {parameters}", query=sql_query_str, parameters=db_documents)
 
@@ -820,64 +780,6 @@ class PgvectorDocumentStore:
 
         return written_docs
 
-    @staticmethod
-    def _from_haystack_to_pg_documents(documents: List[Document]) -> List[Dict[str, Any]]:
-        """
-        Internal method to convert a list of Haystack Documents to a list of dictionaries that can be used to insert
-        documents into the PgvectorDocumentStore.
-        """
-
-        db_documents = []
-        for document in documents:
-            db_document = {k: v for k, v in document.to_dict(flatten=False).items() if k not in ["score", "blob"]}
-
-            blob = document.blob
-            db_document["blob_data"] = blob.data if blob else None
-            db_document["blob_meta"] = Jsonb(blob.meta) if blob and blob.meta else None
-            db_document["blob_mime_type"] = blob.mime_type if blob and blob.mime_type else None
-            db_document["meta"] = Jsonb(db_document["meta"])
-
-            if "sparse_embedding" in db_document:
-                sparse_embedding = db_document.pop("sparse_embedding", None)
-                if sparse_embedding:
-                    logger.warning(
-                        "Document {doc_id} has the `sparse_embedding` field set,"
-                        "but storing sparse embeddings in Pgvector is not currently supported."
-                        "The `sparse_embedding` field will be ignored.",
-                        doc_id=db_document["id"],
-                    )
-
-            db_documents.append(db_document)
-
-        return db_documents
-
-    @staticmethod
-    def _from_pg_to_haystack_documents(documents: List[Dict[str, Any]]) -> List[Document]:
-        """
-        Internal method to convert a list of dictionaries from pgvector to a list of Haystack Documents.
-        """
-
-        haystack_documents = []
-        for document in documents:
-            haystack_dict = dict(document)
-            blob_data = haystack_dict.pop("blob_data")
-            blob_meta = haystack_dict.pop("blob_meta")
-            blob_mime_type = haystack_dict.pop("blob_mime_type")
-
-            # convert the embedding to a list of floats
-            if document.get("embedding") is not None:
-                haystack_dict["embedding"] = document["embedding"].tolist()
-
-            haystack_document = Document.from_dict(haystack_dict)
-
-            if blob_data:
-                blob = ByteStream(data=blob_data, meta=blob_meta, mime_type=blob_mime_type)
-                haystack_document.blob = blob
-
-            haystack_documents.append(haystack_document)
-
-        return haystack_documents
-
     def delete_documents(self, document_ids: List[str]) -> None:
         """
         Deletes documents that match the provided `document_ids` from the document store.
@@ -895,7 +797,7 @@ class PgvectorDocumentStore:
             document_ids_str=SQL(document_ids_str),
         )
 
-        self._ensure_valid_connection()
+        self._ensure_db_setup()
         self._execute_sql(delete_sql, error_msg="Could not delete documents from PgvectorDocumentStore")
 
     async def delete_documents_async(self, document_ids: List[str]) -> None:
@@ -913,8 +815,32 @@ class PgvectorDocumentStore:
             document_ids_str=SQL(document_ids_str),
         )
 
-        await self._ensure_valid_connection_async()
+        await self._ensure_db_setup_async()
         await self._execute_sql_async(delete_sql, error_msg="Could not delete documents from PgvectorDocumentStore")
+
+    def _build_keyword_retrieval_query(self, query: str, top_k: int, filters: Optional[Dict[str, Any]] = None):
+        """
+        Builds the SQL query and the where parameters for keyword retrieval.
+        """
+        sql_select = SQL(KEYWORD_QUERY).format(
+            schema_name=Identifier(self.schema_name),
+            table_name=Identifier(self.table_name),
+            language=SQLLiteral(self.language),
+            query=SQLLiteral(query),
+        )
+
+        where_params = ()
+        sql_where_clause = SQL("")
+        if filters:
+            sql_where_clause, where_params = _convert_filters_to_where_clause_and_params(
+                filters=filters, operator="AND"
+            )
+
+        sql_sort = SQL(" ORDER BY score DESC LIMIT {top_k}").format(top_k=SQLLiteral(top_k))
+
+        sql_query = sql_select + sql_where_clause + sql_sort
+
+        return sql_query, where_params
 
     def _keyword_retrieval(
         self,
@@ -936,25 +862,9 @@ class PgvectorDocumentStore:
             msg = "query must be a non-empty string"
             raise ValueError(msg)
 
-        sql_select = SQL(KEYWORD_QUERY).format(
-            schema_name=Identifier(self.schema_name),
-            table_name=Identifier(self.table_name),
-            language=SQLLiteral(self.language),
-            query=SQLLiteral(query),
-        )
+        sql_query, where_params = self._build_keyword_retrieval_query(query=query, top_k=top_k, filters=filters)
 
-        where_params = ()
-        sql_where_clause = SQL("")
-        if filters:
-            sql_where_clause, where_params = _convert_filters_to_where_clause_and_params(
-                filters=filters, operator="AND"
-            )
-
-        sql_sort = SQL(" ORDER BY score DESC LIMIT {top_k}").format(top_k=SQLLiteral(top_k))
-
-        sql_query = sql_select + sql_where_clause + sql_sort
-
-        self._ensure_valid_connection()
+        self._ensure_db_setup()
         result = self._execute_sql(
             sql_query,
             (query, *where_params),
@@ -963,7 +873,7 @@ class PgvectorDocumentStore:
         )
 
         records = result.fetchall()
-        docs = self._from_pg_to_haystack_documents(records)
+        docs = _from_pg_to_haystack_documents(records)
         return docs
 
     async def _keyword_retrieval_async(
@@ -980,25 +890,9 @@ class PgvectorDocumentStore:
             msg = "query must be a non-empty string"
             raise ValueError(msg)
 
-        sql_select = SQL(KEYWORD_QUERY).format(
-            schema_name=Identifier(self.schema_name),
-            table_name=Identifier(self.table_name),
-            language=SQLLiteral(self.language),
-            query=SQLLiteral(query),
-        )
+        sql_query, where_params = self._build_keyword_retrieval_query(query=query, top_k=top_k, filters=filters)
 
-        where_params = ()
-        sql_where_clause = SQL("")
-        if filters:
-            sql_where_clause, where_params = _convert_filters_to_where_clause_and_params(
-                filters=filters, operator="AND"
-            )
-
-        sql_sort = SQL(" ORDER BY score DESC LIMIT {top_k}").format(top_k=SQLLiteral(top_k))
-
-        sql_query = sql_select + sql_where_clause + sql_sort
-
-        await self._ensure_valid_connection_async()
+        await self._ensure_db_setup_async()
         result = await self._execute_sql_async(
             sql_query,
             (query, *where_params),
@@ -1007,25 +901,18 @@ class PgvectorDocumentStore:
         )
 
         records = await result.fetchall()
-        docs = self._from_pg_to_haystack_documents(records)
+        docs = _from_pg_to_haystack_documents(records)
         return docs
 
-    def _embedding_retrieval(
+    def _check_and_build_embedding_retrieval_query(
         self,
         query_embedding: List[float],
-        *,
+        vector_function: Literal["cosine_similarity", "inner_product", "l2_distance"],
+        top_k: int,
         filters: Optional[Dict[str, Any]] = None,
-        top_k: int = 10,
-        vector_function: Optional[Literal["cosine_similarity", "inner_product", "l2_distance"]] = None,
-    ) -> List[Document]:
+    ):
         """
-        Retrieves documents that are most similar to the query embedding using a vector similarity metric.
-
-        This method is not meant to be part of the public interface of
-        `PgvectorDocumentStore` and it should not be called directly.
-        `PgvectorEmbeddingRetriever` uses this method directly and is the public interface for it.
-
-        :returns: List of Documents that are most similar to `query_embedding`
+        Performs checks and builds the SQL query and the where parameters for embedding retrieval.
         """
 
         if not query_embedding:
@@ -1078,7 +965,30 @@ class PgvectorDocumentStore:
 
         sql_query = sql_select + sql_where_clause + sql_sort
 
-        self._ensure_valid_connection()
+        return sql_query, params
+
+    def _embedding_retrieval(
+        self,
+        query_embedding: List[float],
+        *,
+        filters: Optional[Dict[str, Any]] = None,
+        top_k: int = 10,
+        vector_function: Optional[Literal["cosine_similarity", "inner_product", "l2_distance"]] = None,
+    ) -> List[Document]:
+        """
+        Retrieves documents that are most similar to the query embedding using a vector similarity metric.
+
+        This method is not meant to be part of the public interface of
+        `PgvectorDocumentStore` and it should not be called directly.
+        `PgvectorEmbeddingRetriever` uses this method directly and is the public interface for it.
+
+        :returns: List of Documents that are most similar to `query_embedding`
+        """
+
+        sql_query, params = self._check_and_build_embedding_retrieval_query(
+            query_embedding=query_embedding, vector_function=vector_function, top_k=top_k, filters=filters
+        )
+        self._ensure_db_setup()
         result = self._execute_sql(
             sql_query,
             params,
@@ -1087,7 +997,7 @@ class PgvectorDocumentStore:
         )
 
         records = result.fetchall()
-        docs = self._from_pg_to_haystack_documents(records)
+        docs = _from_pg_to_haystack_documents(records)
         return docs
 
     async def _embedding_retrieval_async(
@@ -1102,53 +1012,12 @@ class PgvectorDocumentStore:
         Asynchronously retrieves documents that are most similar to the query embedding using a
         vector similarity metric.
         """
-        if not query_embedding:
-            msg = "query_embedding must be a non-empty list of floats"
-            raise ValueError(msg)
-        if len(query_embedding) != self.embedding_dimension:
-            msg = (
-                f"query_embedding dimension ({len(query_embedding)}) does not match PgvectorDocumentStore "
-                f"embedding dimension ({self.embedding_dimension})."
-            )
-            raise ValueError(msg)
 
-        vector_function = vector_function or self.vector_function
-        if vector_function not in VALID_VECTOR_FUNCTIONS:
-            msg = f"vector_function must be one of {VALID_VECTOR_FUNCTIONS}, but got {vector_function}"
-            raise ValueError(msg)
-
-        query_embedding_for_postgres = f"'[{','.join(str(el) for el in query_embedding)}]'"
-
-        if vector_function == "cosine_similarity":
-            score_definition = f"1 - (embedding <=> {query_embedding_for_postgres}) AS score"
-        elif vector_function == "inner_product":
-            score_definition = f"(embedding <#> {query_embedding_for_postgres}) * -1 AS score"
-        elif vector_function == "l2_distance":
-            score_definition = f"embedding <-> {query_embedding_for_postgres} AS score"
-
-        sql_select = SQL("SELECT *, {score} FROM {schema_name}.{table_name}").format(
-            schema_name=Identifier(self.schema_name),
-            table_name=Identifier(self.table_name),
-            score=SQL(score_definition),
+        sql_query, params = self._check_and_build_embedding_retrieval_query(
+            query_embedding=query_embedding, vector_function=vector_function, top_k=top_k, filters=filters
         )
 
-        sql_where_clause = SQL("")
-        params = ()
-        if filters:
-            sql_where_clause, params = _convert_filters_to_where_clause_and_params(filters)
-
-        # we always want to return the most similar documents first
-        # so when using l2_distance, the sort order must be ASC
-        sort_order = "ASC" if vector_function == "l2_distance" else "DESC"
-
-        sql_sort = SQL(" ORDER BY score {sort_order} LIMIT {top_k}").format(
-            top_k=SQLLiteral(top_k),
-            sort_order=SQL(sort_order),
-        )
-
-        sql_query = sql_select + sql_where_clause + sql_sort
-
-        await self._ensure_valid_connection_async()
+        await self._ensure_db_setup_async()
         result = await self._execute_sql_async(
             sql_query,
             params,
@@ -1157,5 +1026,5 @@ class PgvectorDocumentStore:
         )
 
         records = await result.fetchall()
-        docs = self._from_pg_to_haystack_documents(records)
+        docs = _from_pg_to_haystack_documents(records)
         return docs
