+++ integrations/pgvector/src/haystack_integrations/document_stores/pgvector/filters.py
@@ -3,7 +3,7 @@
 # SPDX-License-Identifier: Apache-2.0
 from datetime import datetime
 from itertools import chain
-from typing import Any, Dict, List, Literal, Tuple
+from typing import Any, Dict, List, Literal, Optional, Tuple
 
 from haystack.errors import FilterError
 from psycopg.sql import SQL
@@ -19,6 +19,19 @@ PYTHON_TYPES_TO_PG_TYPES = {
 }
 
 NO_VALUE = "no_value"
+
+
+def _validate_filters(filters: Optional[Dict[str, Any]] = None):
+    """
+    Validates the filters provided.
+    """
+    if filters:
+        if not isinstance(filters, dict):
+            msg = "Filters must be a dictionary"
+            raise TypeError(msg)
+        if "operator" not in filters and "conditions" not in filters:
+            msg = "Invalid filter syntax. See https://docs.haystack.deepset.ai/docs/metadata-filtering for details."
+            raise ValueError(msg)
 
 
 def _convert_filters_to_where_clause_and_params(
