+++ integrations/pgvector/tests/conftest.py
@@ -26,7 +26,7 @@ def document_store(request, monkeypatch):
 
     # Ensure connection just for table deletion.
     # During test execution, the tested methods are expected to call _ensure_valid_connection() themselves.
-    store._ensure_valid_connection()
+    store._ensure_db_setup()
     store.delete_table()
 
 
@@ -50,7 +50,7 @@ def document_store_w_hnsw_index(request, monkeypatch):
 
     # Ensure connection just for table deletion.
     # During test execution, the tested methods are expected to call _ensure_valid_connection() themselves.
-    store._ensure_valid_connection()
+    store._ensure_db_setup()
     store.delete_table()
 
 
@@ -59,13 +59,9 @@ def patches_for_unit_tests():
     with patch("haystack_integrations.document_stores.pgvector.document_store.register_vector") as mock_register, patch(
         "haystack_integrations.document_stores.pgvector.document_store.PgvectorDocumentStore.delete_table"
     ) as mock_delete, patch(
-        "haystack_integrations.document_stores.pgvector.document_store.PgvectorDocumentStore._create_table_if_not_exists"
-    ) as mock_create, patch(
-        "haystack_integrations.document_stores.pgvector.document_store.PgvectorDocumentStore._create_keyword_index_if_not_exists"
-    ) as mock_create_kw_index, patch(
         "haystack_integrations.document_stores.pgvector.document_store.PgvectorDocumentStore._handle_hnsw"
     ) as mock_hnsw:
-        yield mock_register, mock_delete, mock_create, mock_create_kw_index, mock_hnsw
+        yield mock_register, mock_delete, mock_hnsw
 
 
 @pytest.fixture
