+++ integrations/pgvector/tests/test_document_store.py
@@ -4,7 +4,6 @@
 
 from unittest.mock import patch
 
-import numpy as np
 import psycopg
 import pytest
 from haystack.dataclasses.document import ByteStream, Document
@@ -33,11 +32,11 @@ class TestDocumentStore(CountDocumentsTest, WriteDocumentsTest, DeleteDocumentsT
         assert retrieved_docs == docs
 
     def test_connection_check_and_recreation(self, document_store: PgvectorDocumentStore):
-        document_store._ensure_valid_connection()
+        document_store._ensure_db_setup()
         original_connection = document_store._connection
 
         with patch.object(PgvectorDocumentStore, "_connection_is_valid", return_value=False):
-            document_store._ensure_valid_connection()
+            document_store._ensure_db_setup()
             new_connection = document_store._connection
 
         # verify that a new connection is created
@@ -50,7 +49,7 @@ class TestDocumentStore(CountDocumentsTest, WriteDocumentsTest, DeleteDocumentsT
 
         # test with new connection
         with patch.object(PgvectorDocumentStore, "_connection_is_valid", return_value=True):
-            document_store._ensure_valid_connection()
+            document_store._ensure_db_setup()
             same_connection = document_store._connection
             assert same_connection is document_store._connection
 
@@ -127,128 +126,6 @@ def test_to_dict(monkeypatch):
     }
 
 
-def test_from_haystack_to_pg_documents():
-    haystack_docs = [
-        Document(
-            id="1",
-            content="This is a text",
-            meta={"meta_key": "meta_value"},
-            embedding=[0.1, 0.2, 0.3],
-            score=0.5,
-        ),
-        Document(
-            id="2",
-            content="This is another text",
-            meta={"meta_key": "meta_value"},
-            embedding=[0.4, 0.5, 0.6],
-            score=0.6,
-        ),
-        Document(
-            id="3",
-            blob=ByteStream(b"test", meta={"blob_meta_key": "blob_meta_value"}, mime_type="mime_type"),
-            meta={"meta_key": "meta_value"},
-            embedding=[0.7, 0.8, 0.9],
-            score=0.7,
-        ),
-    ]
-
-    with patch(
-        "haystack_integrations.document_stores.pgvector.document_store.PgvectorDocumentStore.__init__"
-    ) as mock_init:
-        mock_init.return_value = None
-        ds = PgvectorDocumentStore(connection_string="test")
-
-    pg_docs = ds._from_haystack_to_pg_documents(haystack_docs)
-
-    assert pg_docs[0]["id"] == "1"
-    assert pg_docs[0]["content"] == "This is a text"
-    assert pg_docs[0]["blob_data"] is None
-    assert pg_docs[0]["blob_meta"] is None
-    assert pg_docs[0]["blob_mime_type"] is None
-    assert "dataframe" not in pg_docs[0]
-    assert pg_docs[0]["meta"].obj == {"meta_key": "meta_value"}
-    assert pg_docs[0]["embedding"] == [0.1, 0.2, 0.3]
-    assert "score" not in pg_docs[0]
-
-    assert pg_docs[1]["id"] == "2"
-    assert pg_docs[1]["content"] == "This is another text"
-    assert pg_docs[1]["blob_data"] is None
-    assert pg_docs[1]["blob_meta"] is None
-    assert pg_docs[1]["blob_mime_type"] is None
-    assert "dataframe" not in pg_docs[1]
-    assert pg_docs[1]["meta"].obj == {"meta_key": "meta_value"}
-    assert pg_docs[1]["embedding"] == [0.4, 0.5, 0.6]
-    assert "score" not in pg_docs[1]
-
-    assert pg_docs[2]["id"] == "3"
-    assert pg_docs[2]["content"] is None
-    assert pg_docs[2]["blob_data"] == b"test"
-    assert pg_docs[2]["blob_meta"].obj == {"blob_meta_key": "blob_meta_value"}
-    assert pg_docs[2]["blob_mime_type"] == "mime_type"
-    assert "dataframe" not in pg_docs[2]
-    assert pg_docs[2]["meta"].obj == {"meta_key": "meta_value"}
-    assert pg_docs[2]["embedding"] == [0.7, 0.8, 0.9]
-    assert "score" not in pg_docs[2]
-
-
-def test_from_pg_to_haystack_documents():
-    pg_docs = [
-        {
-            "id": "1",
-            "content": "This is a text",
-            "blob_data": None,
-            "blob_meta": None,
-            "blob_mime_type": None,
-            "meta": {"meta_key": "meta_value"},
-            "embedding": np.array([0.1, 0.2, 0.3]),
-        },
-        {
-            "id": "2",
-            "content": "This is another text",
-            "blob_data": None,
-            "blob_meta": None,
-            "blob_mime_type": None,
-            "meta": {"meta_key": "meta_value"},
-            "embedding": np.array([0.4, 0.5, 0.6]),
-        },
-        {
-            "id": "3",
-            "content": None,
-            "blob_data": b"test",
-            "blob_meta": {"blob_meta_key": "blob_meta_value"},
-            "blob_mime_type": "mime_type",
-            "meta": {"meta_key": "meta_value"},
-            "embedding": np.array([0.7, 0.8, 0.9]),
-        },
-    ]
-
-    ds = PgvectorDocumentStore(connection_string=Secret.from_token("test"))
-    haystack_docs = ds._from_pg_to_haystack_documents(pg_docs)
-
-    assert haystack_docs[0].id == "1"
-    assert haystack_docs[0].content == "This is a text"
-    assert haystack_docs[0].blob is None
-    assert haystack_docs[0].meta == {"meta_key": "meta_value"}
-    assert haystack_docs[0].embedding == [0.1, 0.2, 0.3]
-    assert haystack_docs[0].score is None
-
-    assert haystack_docs[1].id == "2"
-    assert haystack_docs[1].content == "This is another text"
-    assert haystack_docs[1].blob is None
-    assert haystack_docs[1].meta == {"meta_key": "meta_value"}
-    assert haystack_docs[1].embedding == [0.4, 0.5, 0.6]
-    assert haystack_docs[1].score is None
-
-    assert haystack_docs[2].id == "3"
-    assert haystack_docs[2].content is None
-    assert haystack_docs[2].blob.data == b"test"
-    assert haystack_docs[2].blob.meta == {"blob_meta_key": "blob_meta_value"}
-    assert haystack_docs[2].blob.mime_type == "mime_type"
-    assert haystack_docs[2].meta == {"meta_key": "meta_value"}
-    assert haystack_docs[2].embedding == [0.7, 0.8, 0.9]
-    assert haystack_docs[2].score is None
-
-
 @pytest.mark.integration
 def test_hnsw_index_recreation():
     def get_index_oid(document_store, schema_name, index_name):
@@ -266,7 +143,8 @@ def test_hnsw_index_recreation():
     connection_string = "postgresql://postgres:postgres@localhost:5432/postgres"
     schema_name = "test_schema"
     with psycopg.connect(connection_string, autocommit=True) as conn:
-        conn.execute(f"CREATE SCHEMA IF NOT EXISTS {schema_name}")
+        conn.execute(f"DROP SCHEMA IF EXISTS {schema_name} CASCADE")
+        conn.execute(f"CREATE SCHEMA {schema_name}")
 
     # create a first document store and trigger the creation of the hnsw index
     params = {
@@ -276,7 +154,7 @@ def test_hnsw_index_recreation():
         "search_strategy": "hnsw",
     }
     ds1 = PgvectorDocumentStore(**params)
-    ds1._ensure_valid_connection()
+    ds1._ensure_db_setup()
 
     # get the hnsw index oid
     hnws_index_name = "haystack_hnsw_index"
@@ -284,7 +162,7 @@ def test_hnsw_index_recreation():
 
     # create second document store with recreation enabled
     ds2 = PgvectorDocumentStore(**params, hnsw_recreate_index_if_exists=True)
-    ds2._ensure_valid_connection()
+    ds2._ensure_db_setup()
 
     # get the index oid
     second_oid = get_index_oid(ds2, ds2.schema_name, hnws_index_name)
@@ -316,6 +194,7 @@ def test_create_table_if_not_exists():
 
     # Create a new schema
     with psycopg.connect(connection_string, autocommit=True) as conn:
+        conn.execute(f"DROP SCHEMA IF EXISTS {schema_name} CASCADE")
         conn.execute(f"CREATE SCHEMA {schema_name}")
 
     document_store = PgvectorDocumentStore(
@@ -324,17 +203,13 @@ def test_create_table_if_not_exists():
         table_name=table_name,
     )
 
-    document_store._ensure_valid_connection()
-
-    document_store._create_table_if_not_exists()
+    document_store._ensure_db_setup()
 
     first_table_oid = get_table_oid(document_store, schema_name, table_name)
     assert first_table_oid is not None, "Table was not created"
 
-    document_store._create_table_if_not_exists()
-
+    document_store._initialize_table()
     second_table_oid = get_table_oid(document_store, schema_name, table_name)
-
     assert first_table_oid == second_table_oid, "Table was recreated"
 
     # Clean up: drop the schema after the test
